fix(shim): 64-bit FIST_RUNMS watchdog deadline -- 32-bit long overflowed epoch-ms, silently disabling every FBDUMP display flow for ~25-day windows
The FIST_RUNMS dev watchdog (native_main.c, the dump-and-exit seam every display
flow relies on) stored an ABSOLUTE epoch-millisecond deadline (~1.78e12) in a
`long`.  On the -m32 native build (and wasm32) `long` is 32-bit, so
`tv.tv_sec*1000L` overflows: the truncated deadline is NEGATIVE whenever bit 31
of epoch-ms is set, which flips every 2^31 ms (~24.9 days).  In a "negative"
window the `if (g_deadline_ms > 0)` guard fails -> the watchdog never fires ->
the engine stays in its non-returning menu-idle loop -> `timeout` kills the run
(rc=124, no frame).

This is exactly the failure that was surfacing as a load/flaky gate problem:
consecutive.sh iter-1 = 12 passed, 24 FAILED, every FBDUMP display flow
native-rc=124 wasm-rc=124, while the editor-.FSG (roundtrip-exit), intro
(kdvframe-exit) and mission (FIST_MISSFB-exit) flows -- which do NOT use this
watchdog -- passed.  The prior banked 36/36 runs happened to land in a
"positive" epoch window; we are now in a "negative" one.  Root-caused by gdb:
g_deadline_ms = -2135572700 while [0x452] ticks advanced normally (pump running,
watchdog code simply skipped).

Fix: g_deadline_ms and the `now` comparison are long long, with 1000LL / atoll
and an explicit (long long) cast on tv.tv_sec.  native_main.c is the
hand-written platform shim (not the engine decompile), so this is a direct shim
fix, not a patches/NNN diff; the 4 engine md5s stay pristine.

Verified BOTH targets: mainmenu now dumps at exactly 22s (rc=0) on native AND
wasm, native<->wasm 0-diff, AE=0 vs ref/main_menu_native320.png.

+++ tools/native_main.c
@@ -399,13 +399,17 @@ void fist_timer_pump(void){
     /* Dev watchdog: FIST_RUNMS=<ms> dumps the framebuffer (FIST_FBDUMP) and exits after a wall-clock
      * deadline -- lets a first-light frame be captured while the engine is in its (non-returning) main
      * loop.  Runs in main context (safe for the PPM writer). */
-    static long g_deadline_ms = -2;
+    /* NB 64-bit: the deadline is an ABSOLUTE epoch-millisecond value (~1.78e12).  On the -m32 native
+     * build `long` is 32-bit, so tv.tv_sec*1000 overflows it -> the truncated deadline is negative for
+     * ~25-day epoch windows (bit 31 of epoch-ms flips every 2^31 ms), silently disabling this watchdog
+     * (the `> 0` guard below fails) and hanging every FBDUMP display flow until `timeout`.  Use long long. */
+    static long long g_deadline_ms = -2;
     if (g_deadline_ms == -2) {
         const char *r = getenv("FIST_RUNMS");
-        g_deadline_ms = r ? atol(r) : -1;
+        g_deadline_ms = r ? atoll(r) : -1;
         if (g_deadline_ms > 0) {
             struct timeval tv; gettimeofday(&tv, 0);
-            g_deadline_ms += tv.tv_sec*1000L + tv.tv_usec/1000L;
+            g_deadline_ms += (long long)tv.tv_sec*1000LL + tv.tv_usec/1000LL;
         }
     }
     /* Phase-pinned dump (DETERMINISM for tick-phase-dependent screens): FIST_DUMPTICK=N dumps the frame
@@ -425,7 +429,7 @@ void fist_timer_pump(void){
     }
     if (g_deadline_ms > 0) {
         struct timeval tv; gettimeofday(&tv, 0);
-        long now = tv.tv_sec*1000L + tv.tv_usec/1000L;
+        long long now = (long long)tv.tv_sec*1000LL + tv.tv_usec/1000LL;
         if (now >= g_deadline_ms) fist_dump_and_exit("FIST_RUNMS watchdog");
     }
     /* Async vertical-retrace IRQ (palette upload): fires on the timer heartbeat regardless of whether
